Move entrypoint logic and loop management into sdl_app
