+++ src/sf33rd/Source/Game/system/pause.c
@@ -162,13 +162,6 @@ s32 Check_Pause_Term(u16 sw, u8 PL_id) {
         return PAUSE_X = 1;
     }
 
-#if DEBUG
-    // This skips checking controller connection status during gameplay testing
-    if (configuration.test.enabled) {
-        return 0;
-    }
-#endif
-
     if (Present_Mode == 3) {
         if (Interface_Type[Decide_ID] == 0) {
             Pause_ID = Decide_ID;
